+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
--  Support custom environment in e2e tests - [#1645](https://github.com/paritytech/ink/pull/1645)
+-  Support custom environment in E2E tests - [#1645](https://github.com/paritytech/ink/pull/1645)
 
 ### Changed
 - E2E: spawn a separate contracts node instance per test ‒ [#1642](https://github.com/paritytech/ink/pull/1642)